Add state controller
